Start writing the GCP integration
This is pretty rough and the aim is to move most of this functionality
into the `RAP.Storage' module, but while testing it is most convenient
to add it to the `RAP.Job.Producer' module.

There are significant difficulties which arise from the structure of the
GCP storage (it is effectively a flat structure where the web interface's
folder structure is an illusion), as well as needing to check for the
completion and/or running of a job so that we don't endlessly repeat jobs.

FossilOrigin-Name: 9df1ce9826387616f567506ed06ea6a87f2293c9e1559c5d03df55d28fc80257