Start separating the monitoring of storage from the polling of it
The idea here is that the producer stage simply produces descriptions of
jobs which are to be run, and the source of them. This separates regular
polling of various storage backends from the polling proper.

This is advantageous in several respects:

1. Polling just for a list of UUIDs is not expensive computationally and
   and should be separated from the various recursive functions which
   fetch the data proper, which may be computationally expensive and may
   imply race conditions and/or file conflicts.

2. This will put the over-complicated fetching of objects into separate
   module stages which depend on this original producer, and apparently
   to use `ConsumerSupervisor' or similar to parallelise it.

3. The logic which checks that a UUID is not running nor has a cached
   result is really related to producing events, so it's best to locate
   it in the producer stage.

There's a lot involved in separating these, so commit this code as-is
to keep track of changes as I go along. The previous working version did
was not yet joined up with the rest of the stages, anyway, so this is the
ideal time at which to start separating them.

FossilOrigin-Name: 6ae373ad63c5cb255639e68cbeef48ba50195c2fcadde2b38138d4b650f81004