More experiments with monitoring GCP for events
The first thing which is apparent is that events aren't constantly asked
for by `GenStage' consumers/producer-consumers, at least if there are no
events sought. This may be related to the fact that the pipeline isn't
particularly joined up, so there's no back pressure yet.

Regardless of this, there's still the problem of demand. In contrast to
the example in the book, it does seem like handling the requests for
demand is natural, because we effectively have a queue of job UUIDs and
running these is often computationally expensive.

The current system doesn't hold any queue, rather it calls the (now
non-recursive) monitoring function which then produces an arbitrary
number of events.

Thus it seems that it would be better to return to the recursive
monitoring of the GCP storage, and to make good use of the state, which
should store extant UUIDs. This means we can return the right number of
events sought, by using state to hold a queue.

Experimenting with this is productive because it lets us get familiar
with how demand and back-pressure work.

FossilOrigin-Name: 839a034a99328414a8fbaec20c5b08cfe90c6afd48e6f72d36331352da3a9c07