Call the monitoring function(s) with the `Task' module
This ensures that the process continues in the background and keeps
updating state for us.

The shape of this is reasonably good in the sense that events are
produced on-demand. I'm not sure what the implications of the calls are
in terms of events, since they send a reply with an empty list of events.

I'll need to work out how to test that part of the behaviour since the
UUIDs are fairly ephemeral at least in this test phase.

FossilOrigin-Name: 1e428fc359cd70e0907b26ff925b9080c0fa862d74a6d0c5dc1759b290434fd1