Fix object format so that it matches named struct `RAP.Storage'
The previous format was mapping over UUIDs to get a list of objects with
which the UUID is associated.

This format is preferable because it produces a notion of a single object
relating to a job manifest and UUID uploaded to GCP, where there is a
hard-coded `manifest.ttl' file, the resources (data files and schemata),
and the source of the job UUID on GCP.

The `prep_job/2' function, as well as caching the requested UUID in ETS,
now really serves to transform a given UUID, given remote objects grouped
according to UUIDs, into this named `RAP.Storage' struct.

This representation is much more robust and there's much less risk of
fetching/querying objects unneccessarily.

FossilOrigin-Name: bd2282605f581e0021b80be76485445a8b8a5668669c5e418f6d23b38670dfc7