Test the interaction between the producer and a simple, test consumer
Since the jobs are updated very infrequently, the producer is hardly busy
and it appears that the consumers don't keep checking for demand after no
more events are pursued.

This works for now: create one event then adds the rest to the queue.
There's probably a better way to do this (the ``Concurrent Data
Processing in Elixir'' book suggests using Erlang's `:queue'), but this
at least works.

Notably, the example in that book does suggest that the callbacks with
`cast/2', `call/3' &c. are the way to send things upstream quickly, so
perhaps this is actually the best way to advance these messages. It seems
to be independent of the `handle_demand/2' callbacks as described.

FossilOrigin-Name: 6c282bfe24c325d91a92cc305dda9cfef2c3642c16fc95cf050a5ec62f27f9ab