Join up Monitor and GCP stages
This primarily involved rewriting the fetch functions so that they rely
on the session which is renewed every five minutes (if the listing poll
in the Monitor producer module fails). Then, this is basically a series
of pipes and reporting of fetch errors.

Named structs are used where possible so that the pattern matching
catches misuse. Rename the Storage.Transactions module to Storage.Staging
as this describes the function of this module better.

FossilOrigin-Name: 03f9d4c37d5194add43e81d8ec28ba7a3c208e065837aee38bf2a8edeaa8e7a2