Unify MD5 hash check and object fetching function
This prevents the file from being fetched if it already exists on disk.

This can probably be unified into a more terse and clear `with' statement
as currently it combines the `if' macro with subsequent a subsequent
invocation of `with'.

FossilOrigin-Name: 0b13754625353e80ada66dac43b71dbd6d506e47403959707edfcdbec70f2bbc