Use named structs where possible in the `Job.Producer' module
This is desirable because ideally we don't want to change the state of
the Grax description structs so that the fields contain values which
would not be generated by a real RDF graph.

This also nicely carries forward any errors to the end of the stage,
which enables good error reporting.

I returned to the named structs and moved them into new file, `spec.ex'.
This file contains the equivalent of all of the Grax description structs,
and makes clear what is changing and being annotated. The named structs
each require a module (I have nested these) and the @moduledoc pragma
describes the struct and the reasoning behind the fields, and e.g. the
reasoning behind adding two separate table description structs.

The code is also in better shape, especially the map containing valid vs.
invalid columns, which are created with a really succint `call to
`Enum.group_by/2'.

FossilOrigin-Name: ef0a94738dd3d87f34adbde7890c19a5bd4d78d6d2123a9478c4c4f1647259a0