Include resource path in the job column description
This is necessary for pattern-matching over the job structure alone to
decide what jobs are practical.

We actually have access to the tables already, it's just that I'd fed the
names alone into the `Job.Producer.check_column/2' function. Feed in the
valid tables instead, and then extract the path to the table's resource
or data file when checking that the referenced table exists.

Both valid and invalid tables are preserved in the final structure,
however, because this lets us produce informative error messages and
compare structure.

Note, when filtering for table validity, both the existence of the
resource and its corresponding schema are checked. While the program may
be able to run given absence to the schema, this is probably inadvisable
because it implies that there may be some sort of corruption of the data,
or issues uploading all of the dependencies needed for the requested job.
(It is possible to upload data without the schema being valid, but the
validation is done locally using the fish data utilities.)

FossilOrigin-Name: 39a30c318c3f0cf28ad2e90d50f3fd5c203d4329cd18bd93a2312a6f2d988b87