Internal representation should not duplicate cache directory
The cache directory is well-known and hardcoded into each stage's
initial state. The previous internal representation included the UUID and
cache directory in the paths of each object.

This is problematic when moving files at the end into a new directory,
because it means that we sometimes have to remove this leading UUID/cache
directory from the path strings.

It also makes it easier to read the objects when they are printed to logs
and/or returned at the end.

The next step is to nail down the naming of the paths vs file names, as
there isn't a clear difference between these at first glance.f

FossilOrigin-Name: a8b63e2457eb3e14129795d565d6d35bfe9300a0a624d0b296a9ff3585e4c549