Fix-up issues with module structs' attributes
I had a number of spurrious `end' keywords, although the use of atoms
`:start' and `:end' was not the source of errors around too many `end'
invocations, it is better to rename these as it makes the code more
readable.

The code at least compiles, so commit this as-is and see if it runs.

FossilOrigin-Name: 54976401eaf8d0fd3bc53cd2683b7a0efa0ae6a088e3e2be1f5f4d6db37bb82c