Generate HTML documentation
This is rather complicated, but here's a potted summary.

Originally, I had a notion that the generation of HTML documentation
should come *before* caching the result, and that the generation would be
at the end of the pipeline. I think we still want to have the HTML
summary of the manifest processing, at the end of the pipeline, but
caching should come before this. This is advantageous because the
shape/style/contents of the HTML summary is something which may change,
and we don't want to have to re-submit the job just to see this.

Therefore, we ought to cache all of the information we have in the local
Erlang Mnesia database, so the shape of this database table should be the
same as the input to the HTML generation. I had originally added separate
job and manifest tables, but this just complicates querying, and it's
very easy to cache the struct/map as-is.

The way that the HTML documentation generation works is that there are a
bunch of smaller functions which deal with one aspect of the structure,
and return the current working text concatenated with the generated text.
These functions read files in a directory called `./html_fragments' and
use the Elixir EEx ("embedded Elixir") templating library to inject
attributes. The functions are then nicely built up with pipes, which I
like.

Outstanding issues / miscellanea / trivia:

- I've hardcoded a bunch of things in the `RAP.Bakery.Compose' module,
  which should be added to the original %Application{} state fed into
  the module.

- We have YAML as well as TTL schemata, but I need to add these back into
  the structs. It's OK for a  demo so far, though.

- We don't generate descriptive statistics or similar, yet. The point at
  which to generate these isn't exactly clear.

- If we have a way to trigger regeneration of an HTML documentation, we'd
  likely want to create a bunch of processes to do this. Concurrency is
  what Erlang and Elixir are great at. The documentation generation stage
  (the `RAP.Bakery.Compose' module), then, should probably be managed by
  a supervisor, i.e. with `ConsumerSupervisor'.

- The extension is effectively hard-coded as ""json" but it is feasible
  that different types of jobs produce different output data formats.

- We still need to test demand / back-pressure / multiple jobs running at
  once.

- Naming things is very, very hard.

FossilOrigin-Name: aed22718e0afd22118f3a1b2d4d5620b079540d5a71a258ee12f89ac820c1f8c