Add various error 'signals', extract manifest name from index
These changes are largely untested, but they do compile, so check these
in for now.

The basic idea is that errors should not crash the pipeline, since we
need to provide useful feedback. Additionally, which fields are available
depends on the previous state, and we don't want to have a bunch of
different named module structs for the various possible states.

The common signal is `:working' and if this is available, it indicates
that all fields are populated. For most other states, errors are passed
upwards indicating the stage at which [likely] failure occured. Errors
like `:see_runner' are passed upwards, and previous state is recorded
in corresponding struct fields like `:producer_signal'. This is flexible
because it means in principle that succeeding stages can still make use
of what is available up to the that point and record their own failure
state.

Passing failure upwards makes the final output useful and it's more
robust than pattern matching on presence/lack of a field, which could in
principle be used to indicate some of the same issues. We can't let the
pipeline crash in most circumstances.

This also presented a good opportunity in which to get the name of the
manifest object from the index file, as it's closely related to which
fields are available (e.g. if the index file is unreadable, we don't know
the value of this field at all). While this is also in the RDF graph,
similar to how the prefixes are part of the RDF graph, unlike the base
prefix, this is not implicit and we'd have to crawl the RDF graph to
extract it. It's already well-known, so read it from the index file.

As well as clarifying the fields, I created a new module to replace
the `%GCP{}' named struct, since we want to be able to replace the GCP
stage with arbitrary stages which fetch data from various sources.
The idea is that these inject data into this generic struct with common
fields which can be processed regardless of where data were sourced from.

FossilOrigin-Name: 8f44650a2f7982135b66555a4b8e8fa7807ef8119e9129d5f784c1f5b041ee5d