Make sure to test interfacing with the mnesia database
This was really annoying as I hadn't realised that the OTP app needs to
be started. The test_helper file now force-starts the thing and the
simple test of interfacing with the database succeeds.

Further, it reveals that using the Mix environments is the correct way to
have a separate testing and development database.

FossilOrigin-Name: 0ff4cb4e455277fdcedb6792aee8db6149a01d56fceef908db2aac95af00a7ad