Fix-up tests and check pipeline works end-to-end after changes
Most of the issues which arose were due to renaming certain fields,
a good example of which is setting `:output_format' and `:output_stem'
fields on the job result level.

A further element of these changes is to add the full target IRI to the
'index' file (`.index') and read this when loading. This is not the
correct way to do this but it works for now, and it is flexible enough
that it can be substituted for crawling the RDF graph later.

FossilOrigin-Name: 9086aa82a1161fd4cffbe43e4eaf0f6e38b7b1dd903927f3e5b1d69c48aa97f2