Start fleshing out tests for bakery 'compose' module
This currently tests the behaviour for cases where jobs have been run
(signal `:working' or `:job_errors'). There was a further issue with the
time-zone stuff not working, I believe that the fix was to forcibly start
the `:zoneinfo' module, but maybe not, and the fix was unrelated to this.

FossilOrigin-Name: 75bb8f34db4b5bbcb6cca49ba650fdf449d3528a47456b9607d5a2f6e03feca8