Slight changes to the results display
First, don't show a (non-written) result for ignored/invalid jobs, and
fix the pattern-matching for these. The job type is "ignore" but the
signal is the atom `:ignored'.

Second, don't show the title/description for the manifest, at least not
yet. The data model doesn't have these in it for the ManifestDesc object.
We will expand/revise the data model later so that it can output linked
data, and can add this back in when this is implemented.

Third, remove the notion of a source job from the HTML output, since we
don't display the job information yet, just its results.

Finally, add nil case to the `extract_iri/1' `extract_uri/1' functions,
then check for nil when checking if a schema file is extant. This is
really an error-handling case so that a well-described RDF graph which is
uploaded which is missing a schema file name (i.e. conversion to TTL
failed). Again, revisit this when expanding the data model so that the
RDF descriptions are in a better shape.

FossilOrigin-Name: 956bdfcc444998997c816dc40e6836b98f13040f113c65afe9a2c9dc974c8c4d