Rough shape of modules to be serialised to RDF
These are separate from the modules which are the intermediate stage.
The next step will be to merge these with the new modules which are to
be serialised. This also makes sure to sub a fallback 'base' prefix if
reading the index file fails.

FossilOrigin-Name: 41719003131fd0252cb74728b0bd9a51364c2d1223f7d940bdb934ef71418fb7