Pass up state instead of adding more and more signals
For the minimal number of signals we were dealing with (:see_pre,
:see_producer, :see_runner), this was manageable, but it started to
become annoying when serialising TTL, as there wasn't a good reason to
include them in the serialised processed manifest description.

Instead, pass up several notions of state:
1. When the stage was invoked
2. When the stage started the work which produced a given signal
3. When the stage finished the work which produced a given signal

These are currently UNIX timestamps, so need to work out what the correct
data-type for PROV.wasStartedAt (&c.) is.

These are appended to a current :work attribute, and the :signal
attribute is the current signal (which is all we pattern-match on).
In principle, we could pattern match instead on the :work attribute,
which would probably be more robust.

This is much more extensible, and we can set it up for arbitrary stages.
There is one assumption here in that each stage is a single running
process (they aren't started dynamically), so probably include a PID or
similar since we want to track not just the stage name, but the instance
of the stage. This is easy to slot into place now. It's probably a matter
of just updating the function which wraps appending work to past work
passed as an argument.

FossilOrigin-Name: 8ae66a2760d4659559f06a7f4063e515d016d9ff4367c2f305fe773ea3dafe84